pyproject.toml
[project]
name = "factify-ai"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = []

[dependency-groups]
common = [
    "Flask==3.1.2",
    "pymongo==4.15.3",
]

backend = [
    {include-group = "common"},
    "Flask-Cors==3.0.10",
    "python-jose==3.3.0",
    "python-keycloak==3.7.0",
    "pandas==2.0.0",
    "Werkzeug==3.1.3",
    "pypdf==3.1.0",
    "python-docx==0.8.11",
]

cron = [
    {include-group = "common"},
    "pillow==12.0.0",
    "matplotlib==3.7.0",
    "numpy==1.24.0",
    "pandas==2.0.0",
    "scikit-learn==1.3.0",
    "torch==2.2.0",
    "torchvision==0.17.0",
    "tqdm==4.67.1",
    "transformers==4.39.0",
    "google-auth==2.47.0",
    "google-genai==1.60.0",
    "openai==2.15.0",
]
